pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "colext"
version = "0.1.0"
requires-python = ">= 3.8"
description = "FL testbed decorators and launch scripts."
readme = "README.md"
classifiers = [
    "Private :: Do Not Upload"
]

dependencies = [
    "psutil>=5.9.8",
    "psycopg[binary]>=3.1.17",
    "kubernetes>=29.0.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.3",
    "python-on-whales>=0.68.0",
    "flwr>=1.6.0"
]

[project.optional-dependencies]
jetson = ["jetson-stats"]

[project.scripts]
colext_launch_job = "colext.experiments:launch_experiment"
colext_get_metrics = "colext.experiments:retrieve_metrics"

[tool.setuptools_scm]